+++ test/fly_chat_method_channel_test.dart
@@ -6,6 +6,7 @@ import 'package:mirrorfly_plugin/fly_chat_method_channel.dart';
 void main() {
   MethodChannelFlyChatFlutter platform = MethodChannelFlyChatFlutter();
   const MethodChannel channel = MethodChannel('contus.mirrorfly/flyChat');
+  final List<MethodCall> log = <MethodCall>[];
 
   TestWidgetsFlutterBinding.ensureInitialized();
 
@@ -13,12 +14,14 @@ void main() {
     TestDefaultBinaryMessengerBinding.instance?.defaultBinaryMessenger.setMockMethodCallHandler(
       channel,
           (MethodCall methodCall) async {
+        log.add(methodCall);
         return '42';
       },
     );
   });
 
   tearDown(() {
+    log.clear();
     TestDefaultBinaryMessengerBinding.instance?.defaultBinaryMessenger.setMockMethodCallHandler(channel, null);
   });